query-builder/src/main/java/io/twdps/starter/boot/querybuilder/predicate/CollectionContainsPredicate.java
package io.twdps.starter.boot.querybuilder.predicate;

import com.fasterxml.jackson.databind.ObjectMapper;

import java.io.Serializable;

import javax.persistence.criteria.CriteriaBuilder;
import javax.persistence.criteria.CriteriaQuery;
import javax.persistence.criteria.Predicate;
import javax.persistence.criteria.Root;

public class CollectionContainsPredicate<T> extends BasePredicate<T> {

  public CollectionContainsPredicate(
      ObjectMapper objectMapper, Class<? extends Serializable> idClazz) {
    super(objectMapper, idClazz);
  }

  @Override
  public Predicate toPredicate(Root<T> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
    return builder.equal(
        getCriteriaExpressionJoinKey(root),
        parseValue(getCriteriaObjectValue().toString(), builder));
  }
}
